requirements.txt
streamlit
google-generativeai
pandasai
pandas
matplotlib
seaborn
plotly
fastapi==0.115.11
google-generativeai
langchain-google-genai
langchain==0.3.20
langchain-chroma==0.2.2
langchain-community==0.3.19
langchain-core==0.3.43
langchain-elasticsearch==0.3.2
langchain-experimental==0.3.4
langchain-openai==0.3.8
langchain-text-splitters==0.3.6
langgraph==0.2.76
langgraph-checkpoint==2.0.18
langgraph-checkpoint-mongodb==0.1.1
langgraph-checkpoint-redis==0.0.3
langgraph-prebuilt==0.1.2
langgraph-sdk==0.1.55
langsmith==0.3.13
openai==1.65.5
pydantic==2.10.6
pymongo==4.11.2
PyMuPDF==1.25.3
pypdf==5.3.1
python-dotenv==1.0.1
redis==5.2.1
requests==2.32.3
SQLAlchemy==2.0.38
rank_bm25
sse-starlette==2.2.1
uvicorn
beautifulsoup4==4.12.3
sqlite-web
sqlite-utils
aiosqlite
tensorflow
keras
